+++ src/middlewares/index.ts
@@ -1,3 +1,4 @@
 import health from './health';
+import { validateBody, validateQueryString } from './validate-params';
 
-export { health };
+export { health, validateBody, validateQueryString };
