+++ test/initializers/koa/errors/custom-error.test.ts
@@ -1,6 +1,6 @@
 import 'should';
 import * as sinon from 'sinon';
-import { CustomError } from '../../../../src/initializers/koa/errors/custom-error';
+import { CustomError } from '../../../../src/middlewares/errors/custom-error';
 
 const sandbox = sinon.createSandbox();
